# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Login API",
        "version": "0.0.1"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Autenticação do usuário",
                "description": "Autentica um usuário e obter um token de acesso.",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "rhood@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "floresta459"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autenticação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "Token de acesso",
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Senha ou email inválido",
                                            "type": "string",
                                            "example": "Senha ou email inválido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Usuário inativo ou inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Usuário inativo ou inexistente",
                                            "type": "string",
                                            "example": "Usuário inativo ou inexistente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Mensagem de erro do sistema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/me": {
            "post": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Mostra detalhes do usuário autenticado",
                "description": "Retorna detalhes do usuário autenticado.",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida. Retorna os dados atualizados do usuário.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Exibe as informações do usuário autenticado",
                                            "properties": {
                                                "id": {
                                                    "description": "ID do usuário",
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "firstName": {
                                                    "description": "Primeiro nome do usuário",
                                                    "type": "string",
                                                    "example": "Robin"
                                                },
                                                "lastName": {
                                                    "description": "Sobrenome do usuário",
                                                    "type": "string",
                                                    "example": "Hood"
                                                },
                                                "email": {
                                                    "description": "E-mail do usuário",
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "rhood@test.com"
                                                },
                                                "cpf": {
                                                    "description": "CPF do usuário",
                                                    "type": "string",
                                                    "example": "00000000002"
                                                },
                                                "phone": {
                                                    "description": "Número de telefone do usuário",
                                                    "type": "string",
                                                    "example": "81996501010"
                                                },
                                                "status": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "type": {
                                                    "description": "Tipo de usuário",
                                                    "type": "string",
                                                    "example": "user"
                                                },
                                                "email_verified_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": null
                                                },
                                                "password": {
                                                    "description": "Senha do usuário",
                                                    "type": "string",
                                                    "example": "$2y$10$xYPET73j6cGcd2cQGA03DOd3R2AyVwoY3mddzvERatUle1a9KANrO"
                                                },
                                                "created_at": {
                                                    "description": "Data de criação do usuário",
                                                    "type": "string",
                                                    "example": "2023-11-23T10:20:17.000000Z"
                                                },
                                                "updated_at": {
                                                    "description": "Data de atualização do usuário",
                                                    "type": "string",
                                                    "example": "2023-11-23T10:20:17.000000Z"
                                                },
                                                "deleted_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": null
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Mensagem de erro do sistema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Encerra a sessão do usuário",
                "description": "Encerra a sessão do usuário autenticado",
                "operationId": "c4eddaf1a2e990cc90e4326c0d731d41",
                "responses": {
                    "200": {
                        "description": "Sessão encerrada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "Sessão encerrada com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Mensagem de erro do sistema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/refresh": {
            "post": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Atualiza o token de autenticação",
                "description": "Obtém um novo token de acesso com base no token atual.",
                "operationId": "refresh",
                "responses": {
                    "201": {
                        "description": "Token atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Novo Token": {
                                            "description": "Novo token de acesso",
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Mensagem de erro detalhada",
                                            "type": "string",
                                            "example": "Mensagem de erro do sistema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}